requirements.txt
# Python 3.10.11
h5py==3.10.0
matplotlib==3.8.3
mtcnn==0.1.1
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
python-dotenv==1.0.1
scikit-learn==1.4.1.post1
tensorflow==2.15.0.post1